Move images to devfile directory
